refactor: 移动模块到核心目录并更新导入路径
