+++ App/src/components/setting/PluginSetting.vue
@@ -3,7 +3,7 @@ import {useApplicationStore} from "../../stores/useApplicationStore";
 import {officialPlugins} from "../../plugins/officialPlugins";
 import TcSwitch from "../../UI/TcSwitch.vue";
 import TcButton from "../../UI/TcButton.vue";
-import {init_plugins} from "../../modules/pluginsManager";
+import {init_plugins} from "../../core/plugins-systems/pluginsManager";
 // 数据流动: 按钮更新 --> store变化 --> 按钮状态变化
 const store = useApplicationStore();
