+++ App/src/plugins/scheduleEditor/store/scheduleStore.ts
@@ -1,6 +1,6 @@
 import {ScheduleEditorProfileStore} from "../scheduleEditor.types";
 import {PluginStore} from "../../../types/plugins.types";
-import {PluginFs} from "../../../modules/pluginUtils";
+import {PluginFs} from "../../../core/utils/pluginUtils";
 import {clearProfile, initProfile, scheduleEditorProfile} from "./scheduleEditorProfile";
 import {clearTodayConfig, initTodayConfig} from "./todayConfigStore";
